chore(nextjs): Extract env variables to separate module and drop their export from public API
