+++ packages/nextjs/src/server/constants.ts
@@ -1,0 +1,17 @@
+/**
+ * @deprecated
+ */
+export const JS_VERSION = process.env.CLERK_JS_VERSION || '';
+export const CLERK_JS_VERSION = process.env.NEXT_PUBLIC_CLERK_JS_VERSION || '';
+export const CLERK_JS_URL = process.env.NEXT_PUBLIC_CLERK_JS || '';
+export const API_URL = process.env.CLERK_API_URL || 'https://api.clerk.dev';
+export const API_VERSION = process.env.CLERK_API_VERSION || 'v1';
+export const API_KEY = process.env.CLERK_API_KEY || '';
+export const SECRET_KEY = process.env.CLERK_SECRET_KEY || '';
+export const FRONTEND_API = process.env.NEXT_PUBLIC_CLERK_FRONTEND_API || '';
+export const PUBLISHABLE_KEY = process.env.NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY || '';
+export const DOMAIN = process.env.NEXT_PUBLIC_CLERK_DOMAIN || '';
+export const PROXY_URL = process.env.NEXT_PUBLIC_CLERK_PROXY_URL || '';
+export const IS_SATELLITE = process.env.NEXT_PUBLIC_CLERK_IS_SATELLITE === 'true' || false;
+export const SIGN_IN_URL = process.env.NEXT_PUBLIC_CLERK_SIGN_IN_URL || '';
+export const SIGN_UP_URL = process.env.NEXT_PUBLIC_CLERK_SIGN_UP_URL || '';
