+++ packages/nextjs/src/server/getAuth.ts
@@ -11,7 +11,7 @@ import {
 import type { SecretKeyOrApiKey } from '@clerk/types';
 
 import { withLogger } from '../utils/debugLogger';
-import { API_KEY, API_URL, API_VERSION, SECRET_KEY } from './clerkClient';
+import { API_KEY, API_URL, API_VERSION, SECRET_KEY } from './constants';
 import { getAuthAuthHeaderMissing } from './errors';
 import type { RequestLike } from './types';
 import { getAuthKeyFromRequest, getCookie, getHeader, injectSSRStateIntoObject } from './utils';
