+++ packages/nextjs/src/server/utils.ts
@@ -4,7 +4,7 @@ import type { NextRequest } from 'next/server';
 import { NextResponse } from 'next/server';
 
 import { constants as nextConstants } from '../constants';
-import { API_KEY, DOMAIN, IS_SATELLITE, PROXY_URL, SECRET_KEY, SIGN_IN_URL } from './clerkClient';
+import { API_KEY, DOMAIN, IS_SATELLITE, PROXY_URL, SECRET_KEY, SIGN_IN_URL } from './constants';
 import { missingDomainAndProxy, missingSignInUrlInDev } from './errors';
 import type { NextMiddlewareResult, RequestLike } from './types';
 import type { WithAuthOptions } from './types';
